+++ src/layouts/Layout_02.jsx
@@ -11,33 +11,38 @@ import { languageLevels, skillLevels } from "../data/textLevelTemplates";
 
 export default function Layout_02({data})
 {
-    const useDarkText = isBright(data.accentColor);
-    const textColor = useDarkText ? "var(--black)" : "var(--white)";
-    const iconColor = useDarkText ? "invert(0)" : "invert(1)";
+    const brightAccent = isBright(data.accentColor);
+
+    const cssVariables = {
+        "--background": data.accentColor,
+        "--text-color": brightAccent ? "var(--black)" : "var(--white)",
+        "--icon-filter": brightAccent ? "invert(0)" : "invert(1)",
+        "--font": data.font
+    };
 
     return (
-        <div style={{fontFamily: data.font}} className={styles["layout-02"]}>
+        <div style={cssVariables} className={styles["layout-02"]}>
             <div className={styles["left"]}>
-                <div style={{backgroundColor: data.accentColor}} className={styles["header-wrapper"]}>
-                    <h1 style={{color: textColor}} className={styles["name"]}>{data.personalInfo.fullName}</h1>
-                    <h2 style={{color: textColor}} className={styles["profession"]}>{data.personalInfo.profession}</h2>
-                    <div style={{borderBottom: `solid calc(var(--width) * .002) ${textColor}`}} className={styles["seperator"]}></div>
+                <div className={styles["header-wrapper"]}>
+                    <h1 className={styles["name"]}>{data.personalInfo.fullName}</h1>
+                    <h2 className={styles["profession"]}>{data.personalInfo.profession}</h2>
+                    <div className={styles["seperator"]}></div>
                     <div className={styles["header-items-wrapper"]}>
                         {
                             data.personalInfo.address !== "" &&
-                            <HeaderItem textColor={textColor} iconColor={iconColor} icon={appIcons.address} text={data.personalInfo.address}/>
+                            <HeaderItem icon={appIcons.address} text={data.personalInfo.address}/>
                         }
                         {
                             data.contact.phoneNumber !== "" &&
-                            <HeaderItem textColor={textColor} iconColor={iconColor} icon={appIcons.contact} text={data.contact.phoneNumber}/>
+                            <HeaderItem icon={appIcons.contact} text={data.contact.phoneNumber}/>
                         }
                         {
                             data.contact.email !== "" &&
-                            <HeaderItem textColor={textColor} iconColor={iconColor} icon={appIcons.email} text={data.contact.email}/>
+                            <HeaderItem icon={appIcons.email} text={data.contact.email}/>
                         }
                         {
                             data.personalInfo.zip !== "" &&
-                            <HeaderItem textColor={textColor} iconColor={iconColor} icon={appIcons.zip} text={data.personalInfo.zip}/>
+                            <HeaderItem icon={appIcons.zip} text={data.personalInfo.zip}/>
                         }
                         {
                             data.links.map(item => {
@@ -46,8 +51,6 @@ export default function Layout_02({data})
                                     !item.hidden &&
                                     <HeaderItem
                                         key={item.id}
-                                        textColor={textColor}
-                                        iconColor={iconColor}
                                         icon={linkIcons.find(icon => icon.name === item.icon).icon}
                                         text={item.url}
                                     />
@@ -56,39 +59,39 @@ export default function Layout_02({data})
                         }
                     </div>
                 </div>
-                    <LanguagesSection accentColor={data.accentColor} iconColor={iconColor} data={data}/>
+                    <LanguagesSection data={data}/>
             </div>
             <div className={styles["right"]}>
                 {
                     data.order === "educationFirst" ? (
                         <>
-                            <EducationSection accentColor={data.accentColor} iconColor={iconColor} educationITems={data.education}/>
-                            <ExperienceSection accentColor={data.accentColor} iconColor={iconColor} experienceItems={data.experience}/>
+                            <EducationSection educationITems={data.education}/>
+                            <ExperienceSection experienceItems={data.experience}/>
                         </>
                     ) : (
                         <>
-                            <ExperienceSection accentColor={data.accentColor} iconColor={iconColor} experienceItems={data.experience}/>
-                            <EducationSection accentColor={data.accentColor} iconColor={iconColor} educationITems={data.education}/>
+                            <ExperienceSection experienceItems={data.experience}/>
+                            <EducationSection educationITems={data.education}/>
                         </>
                     )
                 }
-                <SkillsSection accentColor={data.accentColor} iconColor={iconColor} data={data}/>
+                <SkillsSection data={data}/>
             </div>
         </div>
     )
 
-    function SkillsSection({accentColor, iconColor, data})
+    function SkillsSection({data})
     {
         return (
             !isEmptySection(data.skills) &&
             <div className={styles["section-wrapper"]}>
                 <h3 className={styles["section-title"]}>
-                    <div className={styles["section-icon-wrapper"]} style={{backgroundColor: accentColor}}>
-                        <img style={{filter: iconColor}} className={styles["section-icon"]} src={appIcons.skills} alt="" />
+                    <div className={styles["section-icon-wrapper"]}>
+                        <img className={styles["section-icon"]} src={appIcons.skills} alt="" />
                     </div>
                     Skills
                 </h3>
-                <div className={styles["seperator"]}></div>
+                <div className={styles["seperator-body"]}></div>
                 <div className={styles["flow-wrapper"]}>
                 {
                     data.skills.map(item => {
@@ -102,7 +105,7 @@ export default function Layout_02({data})
                                     <>
                                         {
                                             data.levelMode === "bar" &&
-                                            <LevelBar level={item.level} accentColor={accentColor}/>
+                                            <LevelBar level={item.level}/>
                                         }
                                         {
                                             data.levelMode === "text" &&
@@ -123,18 +126,18 @@ export default function Layout_02({data})
         )
     }
 
-    function ExperienceSection({accentColor, iconColor, experienceItems})
+    function ExperienceSection({experienceItems})
     {
         return (
             !isEmptySection(experienceItems) &&
             <div className={styles["section-wrapper"]}>
                 <h3 className={styles["section-title"]}>
-                    <div className={styles["section-icon-wrapper"]} style={{backgroundColor: accentColor}}>
-                        <img style={{filter: iconColor}} className={styles["section-icon"]} src={appIcons.experience} alt="" />
+                    <div className={styles["section-icon-wrapper"]}>
+                        <img className={styles["section-icon"]} src={appIcons.experience} alt="" />
                     </div>
                     Experience
                 </h3>
-                <div className={styles["seperator"]}></div>
+                <div className={styles["seperator-body"]}></div>
                 {
                     experienceItems.map(item => {
                         return(
@@ -168,18 +171,18 @@ export default function Layout_02({data})
         )
     }
 
-    function EducationSection({accentColor, iconColor, educationITems})
+    function EducationSection({educationITems})
     {
         return (
             !isEmptySection(educationITems) &&
             <div className={styles["section-wrapper"]}>
                 <h3 className={styles["section-title"]}>
-                    <div className={styles["section-icon-wrapper"]} style={{backgroundColor: accentColor}}>
-                        <img style={{filter: iconColor}} className={styles["section-icon"]} src={appIcons.education} alt="" />
+                    <div className={styles["section-icon-wrapper"]}>
+                        <img className={styles["section-icon"]} src={appIcons.education} alt="" />
                     </div>
                     Education
                 </h3>
-                <div className={styles["seperator"]}></div>
+                <div className={styles["seperator-body"]}></div>
                 {
                     educationITems.map(item => {
                         return(
@@ -211,18 +214,18 @@ export default function Layout_02({data})
         )
     }
 
-    function LanguagesSection({accentColor, iconColor, data})
+    function LanguagesSection({data})
     {
         return (
             !isEmptySection(data.languages) &&
             <div className={styles["section-wrapper"]}>
                 <h3 className={styles["section-title"]}>
-                    <div className={styles["section-icon-wrapper"]} style={{backgroundColor: accentColor}}>
-                        <img style={{filter: iconColor}} className={styles["section-icon"]} src={appIcons.languages} alt="" />
+                    <div className={styles["section-icon-wrapper"]}>
+                        <img className={styles["section-icon"]} src={appIcons.languages} alt="" />
                     </div>
                     Languages
                 </h3>
-                <div className={styles["seperator"]}></div>
+                <div className={styles["seperator-body"]}></div>
                 <div className={styles["flow-wrapper"]}>
                 {
                     data.languages.map(item => {
@@ -236,7 +239,7 @@ export default function Layout_02({data})
                                     <>
                                         {
                                             data.levelMode === "bar" &&
-                                            <LevelBar level={item.level} accentColor={accentColor}/>
+                                            <LevelBar level={item.level}/>
                                         }
                                         {
                                             data.levelMode === "text" &&
@@ -257,21 +260,21 @@ export default function Layout_02({data})
         )
     }
 
-    function LevelBar({level, accentColor})
+    function LevelBar({level})
 {
     return (
         <div className={styles["level-container"]}>
-            <span style={{width: level + "%", backgroundColor: accentColor}} className={styles["level"]}></span>
+            <span style={{"--level-width": `${level}%`}} className={styles["level"]}></span>
         </div>
     )
 }
 
-    function HeaderItem({textColor, iconColor, text, icon})
+    function HeaderItem({text, icon})
     {
         return (
             <div className={styles["header-item"]}>
-                <img style={{filter: iconColor}} className={styles["header-item-icon" ]}src={icon} alt="" />
-                <p style={{color: textColor}} className={styles["header-item-text"]}>{text}</p>
+                <img className={styles["header-item-icon" ]}src={icon} alt="" />
+                <p className={styles["header-item-text"]}>{text}</p>
             </div>
         )
     }